requirements.txt
kafka-python==2.0.2
fastapi==0.115.1
uvicorn==0.15.0
pydantic==2.9.2
asyncio==3.4.3
